requirements.txt
# Core
Django>=4.2,<5.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3

# API schema (drf-spectacular replaces drf-yasg)
drf-spectacular>=0.27

# Filtering & pagination
django-filter>=23.5

# Database
mysqlclient>=2.2

# File / object storage
boto3>=1.34
django-storages>=1.14
Pillow>=10.0

# Email (Resend)
resend>=0.7

# Payments
razorpay>=1.4

# Environment
python-dotenv>=1.0

# Server
gunicorn>=21.2

# PDF generation
reportlab>=4.0

# Utilities
python-dateutil>=2.8.2,<3.0.0
